move ti files out of applicant view
