+++ server/app/views/applicant/ProgramOverviewView.java
@@ -21,7 +21,6 @@ import services.applicant.ApplicantPersonalInfo;
 import services.applicant.ApplicantService;
 import services.program.ProgramDefinition;
 import services.settings.SettingsManifest;
-import views.ApplicantBaseView;
 import views.components.TextFormatter;
 
 /**
