+++ server/app/views/errors/NotFound.java
@@ -11,7 +11,7 @@ import services.BundledAssetsFinder;
 import services.DeploymentType;
 import services.applicant.ApplicantPersonalInfo;
 import services.settings.SettingsManifest;
-import views.ApplicantBaseView;
+import views.applicant.ApplicantBaseView;
 
 /**
  * Renders a page to handle 404 not found errors that will be shown to users instead of the unthemed
