+++ server/app/views/support/UnsupportedBrowserView.java
@@ -12,8 +12,8 @@ import play.mvc.Http;
 import play.twirl.api.Content;
 import views.BaseHtmlView;
 import views.HtmlBundle;
-import views.applicant.ApplicantLayout;
 import views.style.StyleUtils;
+import views.trustedintermediary.ApplicantLayout;
 
 /** View responsible for rendering pages showing information regarding unsupported browsers. */
 public final class UnsupportedBrowserView extends BaseHtmlView {
